+++ packages/react-native-callingx/src/CallingxModule.ts
@@ -89,6 +89,7 @@ class CallingxModule implements ICallingxModule {
         incomingChannel,
         ongoingChannel,
         notificationTexts,
+        skipIncomingPushInForeground = false,
       } = options.android ?? {};
 
       this.titleTransformer =
@@ -104,6 +105,7 @@ class CallingxModule implements ICallingxModule {
           ...(ongoingChannel ?? {}),
         },
         notificationTexts,
+        skipIncomingPushInForeground,
       };
 
       if (